BLDC with hall sensors running inside DMA1 Channel1 interrupt, and seeing values returned from the DMA. Instruction stepping the interrupt suggests that the current program easily fits inside the interrupt time, but there are a lot of expensive overheads to be trimmed - noteably the GPIO reads, and swapping between functions. Comparator safety functions still work fine. Opamps still broken, but might be hardware. Motor runs hot, and from the oscilloscope traces it seems motor under test might have erroneous hall phase; it seems the current strongly ramps up towards the end of each BLDC step

+++ Core/Inc/MESCfoc.h
@@ -35,7 +35,7 @@ void foc_init( MESCfoc_s * foc );
 
 typedef struct
 {
-    uint16_t RawADC[FOC_NUM_ADC][FOC_CONV_CHANNELS]; 	//ADC1 returns Ucurrent, DClink voltage and U phase voltage
+    uint32_t RawADC[FOC_NUM_ADC][FOC_CONV_CHANNELS]; 	//ADC1 returns Ucurrent, DClink voltage and U phase voltage
     													//ADC2 returns Vcurrent, V and W phase voltages
 														//ADC3 returns Wcurrent,
     uint16_t ADCOffset[FOC_NUM_ADC];	//During detect phase, need to sense the zero current offset
